controle des session

+++ appli/controller/products/delete_product.php
@@ -1,6 +1,11 @@
 <?php
 use appli\repository\ProductsRepository;
 
+if (!$oUserSession->isConnected() || $oUserSession->getRole() != 'admin') {
+    header('Location:?action=login');
+    exit;
+}
+
 $id = $_GET['id'];
 $oPdo = new ProductsRepository();
 $oPdo->deleteProduct($id);
